requirements.txt
# requirements.txt
python-chess==1.999
numpy>=1.20.0
Pillow>=8.0.0
